+++ src/contentscript/contexts/mutable-web-context/mutable-web-provider.tsx
@@ -34,10 +34,6 @@ const MutableWebProvider: FC<Props> = ({ children, engine }) => {
   )
 
   useEffect(() => {
-    setSelectedMutationId(selectedMutation?.id ?? null)
-  }, [selectedMutation])
-
-  useEffect(() => {
     loadMutations(engine)
   }, [engine])
