+++ src/contentscript/multitable-panel/components/image.tsx
@@ -13,13 +13,14 @@ export interface Props {
 export const Image: FC<Props> = ({ image, alt, fallbackUrl }) => {
   const [imageUrl, setImageUrl] = useState<string | undefined>(undefined)
 
+  // todo: image can changed. need watch
   useEffect(() => {
     image?.ipfs_cid
       ? setImageUrl(`https://ipfs.near.social/ipfs/${image.ipfs_cid}`)
       : image?.url
       ? setImageUrl(image?.url)
       : setImageUrl(fallbackUrl)
-  }, [image, imageUrl])
+  }, [image])
 
   return (
     <img
